Update index.htm
Updated the wording on the ghoul trap and added the randomly generated starting room. Prior to this, the start room was always the same, with the stairs in the middle and a door on each side. Now it has a random stair location and a random number of doors, 1-4, along any of the sides.