Update index.htm
Added a 15% chance it would reduce the number of doors in a room by one. Also set it so that it won't add a door into a room that is already open, since the odds are you would have already seen the door if it was there. On top of this, I also added an arrow to the stairs tile, telling you which direction the stairs now point. Added a check to make sure it wouldn't add a door at the end of the halls where they meet the middle room, though I'm pretty sure this is rendered moot by no longer adding a door to an opened room, but I did this first.