Update index.htm
Reduced the odds of getting a search item by over half. We were getting way too many in testing this. Also reduced the odds of a block in a hall just slightly. Reworded the extra rules section to show that you now click on a room, not hit the open room button. That was from pre-graphical version of this program. Also added some different wording for how thrown weapons work for the wizard.