Update index.htm
Added a function that opens the other half of an outer hallway if you click on one. For example. the top hall-- if you click on left side, it will open both left and right sides of that hallway. 

Added a specific rule for how to deal with Escape spells cast by a creature.

Fixed a bug in level 20 and above where it didn't list the proper gold amount for reward on end room.