Update index.htm
Increased odds of getting a fire demon significantly, especially as the level increases. It was nearly impossible before.

Also increased odds of getting an artifact by quite a bit. I lowered the odds before, which was fine, but then I lowered the overall chance of getting a search item, which made the odds of any of those being an artifact very low. I increased it so we should occasionally get one.