Update of index to fix a couple of bugs
The room with a notch out of the corner would sometimes show what was in that corner in the other room, if there was anything. I also noticed a bug in the notes below when displaying STb and Thr, it would just blank them. I fixed that (with a lot of effort!). I also fixed a problem with undo not reloading the old notes.