Update index.htm
There was a little piece of code in the function roomsShareADoor that checks that corner where hall 32 meets 36 and 27 (top, left, right around center room), and it requires room1 to be less than room2. Same for 33 meeting 36 & 37. Depending on the order the halls were opened in, this could sometimes allow a door in the corner on a room that already had a door into the hall along that wall. This should fix that. It's labeled 6-15-25.