Update index.htm
Maded modifications (tagged with 8-24-24) to the getMonster and openRoomVisually functions to look for "Observe Max" in special section of creatureList. This was so we can set an absolute max number of a specific creature per quest, such as only one of a creature ever showing up. Before this we limited a max per room, not quest.